requirements.txt
# Flask API
flask
flask-cors

# Database
pymongo

# PDF processing
reportlab
PyPDF2

# Environment variables
python-dotenv

# For CSV / JSON (standard lib already has support but keep compatibility)
pandas
